requirements.txt
HAP-python==4.9.*
cryptography==41.0.*
nfcpy==1.0.*
cbor2==5.4.*
pyusb==1.2.*
requests==2.32.*
websockets==14.2
paho-mqtt==2.1.*
